+++ waywiser_carla/launch/carla.launch.py
@@ -5,7 +5,6 @@ from launch import LaunchDescription
 from launch.actions import DeclareLaunchArgument
 from launch.actions import IncludeLaunchDescription
 from launch.actions import OpaqueFunction
-from launch.conditions import IfCondition
 from launch.launch_description_sources import PythonLaunchDescriptionSource
 from launch.substitutions import LaunchConfiguration
 from launch_ros.actions import Node
@@ -27,9 +26,6 @@ def generate_launch_description():
 
     # use context to start nodes
     carla_orchestrator_launch_action = OpaqueFunction(function=carla_orchestrator_launch)
-    emulated_angle_sensor_conditional_launch_action = OpaqueFunction(
-        function=emulated_angle_sensor_conditional_launch
-    )
     rgbd_to_pointcloud_launch_action = OpaqueFunction(function=rgbd_to_pointcloud_launch)
 
     # create launch description
@@ -41,7 +37,6 @@ def generate_launch_description():
 
     # start nodes
     ld.add_action(carla_orchestrator_launch_action)
-    ld.add_action(emulated_angle_sensor_conditional_launch_action)
     ld.add_action(rgbd_to_pointcloud_launch_action)
 
     return ld
@@ -102,29 +97,6 @@ def carla_orchestrator_launch(context):
     )
 
     return [carla_orchestrator]
-
-
-def emulated_angle_sensor_conditional_launch(context):
-    enable_emulated_angle_sensor = False
-    with open(LaunchConfiguration('carla_config').perform(context)) as f:
-        config_data = yaml.safe_load(f)
-        emulated_angle_sensor_params_dict = config_data['emulated_angle_sensor']['ros__parameters']
-        if 'enable' in emulated_angle_sensor_params_dict:
-            enable_emulated_angle_sensor = emulated_angle_sensor_params_dict['enable']
-
-    emulated_angle_sensor_node = Node(
-        package='waywiser_carla',
-        executable='emulated_angle_sensor',
-        name='emulated_angle_sensor',
-        parameters=[
-            {'use_sim_time': LaunchConfiguration('use_sim_time')},
-            LaunchConfiguration('carla_config'),
-        ],
-        output='screen',
-        condition=IfCondition(str(enable_emulated_angle_sensor)),
-    )
-
-    return [emulated_angle_sensor_node]
 
 
 def rgbd_to_pointcloud_launch(context):
